# Optimise getFilteredUTxO