In my microbench (100,000-entry UTxO with 100 different addresses, 50 of those
are in the Set), `getFilteredUTxO` is now 8x faster.

Before:

    time                 167.2 ms   (160.2 ms .. 173.7 ms)
                         0.997 R²   (0.986 R² .. 1.000 R²)
    mean                 170.4 ms   (167.0 ms .. 175.9 ms)
    std dev              6.315 ms   (4.523 ms .. 9.075 ms)
    variance introduced by outliers: 12% (moderately inflated)

After:

    time                 18.99 ms   (18.37 ms .. 19.95 ms)
                         0.991 R²   (0.982 R² .. 0.997 R²)
    mean                 18.74 ms   (18.34 ms .. 19.18 ms)
    std dev              1.006 ms   (799.3 μs .. 1.316 ms)
    variance introduced by outliers: 21% (moderately inflated)